+++ src/pan3d/viewer/__init__.py
@@ -1,5 +1,5 @@
-from .main import main
+from .__main__ import main
 
 __all__ = [
-    "main",
+    "__main__",
 ]
