+++ src/tests/test_import.py
@@ -1,2 +1,2 @@
 def test_import():
-    from pan3d.viewer import main  # noqa: F401
+    from pan3d.viewer.__main__ import main  # noqa: F401
